Fix parsing of consecutive comments
This commit fixes a bug where a comment without a discussion part
followed by another comment would incorrectly parse as a single comment.

We use a negative lookahead to make sure that the discussion part of a
comment does not start with the comment prefix.

See also: https://stackoverflow.com/a/406408/11494525

```
nitpick (non-blocking): This is not worded correctly.

question (non-blocking): At this point, does it matter which thread has won?

Maybe to prevent a race condition we should keep looping until they've all won?
```

The text above would incorrectly parse as:

```
{
    "label": "nitpick",
    "subject": "This is not worded correctly.",
    "decoration": ["non-blocking"],
    "discussion": "question (non-blocking): At this point, does it matter which thread has won?\n\nMaybe to prevent a race condition we should keep looping until they've all won?"
}
```