Add `node_modules` to .gitignore
Even though I have an entry for `node_modules` in my global `~/.gitignore`,
it probably makes sense to add it on the project level too.

While I am a proponent of putting as much stuff as possible in a global
`~/.gitignore` file, I can't reasonably expect every contributor to have the same setup as me.
Thus, by adding trivial entries such as this to a local `.gitignore` for every project,
I hope to lower the barrier to contributing.

Additionally, there is a bug in [Atom](https://atom.io), where the global `~/.gitignore`
is ignored if there is a `.gitignore` file present in the project directory that could
cause files to inadvertently be commited when using Atom's VCS support.